requirements.txt

fastapi==0.110.0
llama_index==0.9.28.post2
openai==1.12.0
pydantic==2.6.3
python-dotenv==1.0.1
SQLAlchemy==2.0.25
supabase==2.3.8
chromadb==0.4.16